@@ -12,44 +12,44 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "6.0.1",
-    "@angular/cdk": "6.0.1",
-    "@angular/common": "6.0.1",
-    "@angular/compiler": "6.0.1",
-    "@angular/core": "6.0.1",
-    "@angular/flex-layout": "5.0.0-beta.14",
-    "@angular/forms": "6.0.1",
-    "@angular/http": "6.0.1",
-    "@angular/material": "6.0.1",
-    "@angular/platform-browser": "6.0.1",
-    "@angular/platform-browser-dynamic": "6.0.1",
-    "@angular/router": "6.0.1",
-    "@angular/service-worker": "6.0.1",
-    "core-js": "2.5.6",
-    "immutable": "3.8.2",
-    "rxjs": "6.1.0",
-    "rxjs-compat": "^6.1.0",
-    "zone.js": "0.8.26"
+    "@angular/animations": "^7.2.15",
+    "@angular/cdk": "^7.3.7",
+    "@angular/common": "^7.2.15",
+    "@angular/compiler": "^7.2.15",
+    "@angular/core": "^7.2.15",
+    "@angular/flex-layout": "^7.0.0-beta.24",
+    "@angular/forms": "^7.2.15",
+    "@angular/http": "^7.2.15",
+    "@angular/material": "^7.3.7",
+    "@angular/platform-browser": "^7.2.15",
+    "@angular/platform-browser-dynamic": "^7.2.15",
+    "@angular/router": "^7.2.15",
+    "@angular/service-worker": "^7.2.15",
+    "core-js": "^2.5.6",
+    "immutable": "^3.8.2",
+    "rxjs": "^6.1.0",
+    "tslib": "^1.9.0",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "6.0.1",
-    "@angular/compiler-cli": "6.0.1",
-    "@angular/language-service": "6.0.1",
-    "@types/jasmine": "2.8.7",
-    "@types/jasminewd2": "2.0.3",
-    "@types/node": "6.0.109",
-    "codelyzer": "4.3.0",
-    "jasmine-core": "2.8.0",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "2.0.2",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-jasmine": "1.1.2",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "protractor": "5.1.2",
-    "ts-node": "4.1.0",
-    "tslint": "5.9.1",
-    "typescript": "2.7.2",
-    "@angular-devkit/build-angular": "0.6.1"
+    "@angular/cli": "^7.3.9",
+    "@angular/compiler-cli": "^7.2.15",
+    "@angular/language-service": "^7.2.15",
+    "@types/jasmine": "^2.8.7",
+    "@types/jasminewd2": "^2.0.3",
+    "@types/node": "^6.0.109",
+    "codelyzer": "^4.3.0",
+    "jasmine-core": "^2.8.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^2.0.2",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-coverage-istanbul-reporter": "^1.4.2",
+    "karma-jasmine": "^1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "^5.1.2",
+    "ts-node": "^4.1.0",
+    "tslint": "^5.9.1",
+    "typescript": ">=3.1.1 <3.3",
+    "@angular-devkit/build-angular": "~0.13.0"
   }
 }
